seperating ASR generation and SLTev

+++ src/automate_ASR.sh
@@ -1,6 +1,12 @@
-#!/bin/bash
-# source bashelan.sh
+!/bin/bash
 source common.sh
+
+if [ $# -ne 2 ] ; then
+    echo "usage: $0 <path_to_index> <worker_fp>"
+    exit
+fi
+
+
 echo "Cleaning the directory and copying new files"
 mkdir -p ../../../SLTev/SLTev-cache/OStt-tt-files/
 rm -rf ../../../SLTev/SLTev-cache/OStt-tt-files/*
@@ -32,12 +38,15 @@ do
 	ffmpeg -hide_banner -loglevel panic -y -i $i -acodec pcm_s16le -ar 16000 -ac 1 $op_wav
 done
 
-# Processing ASR of files
+Processing ASR of files
 
 FROMASR=$2
+#Uncomment it after this run
+#FROM=$3
 index_name=$(basename $1)
-
+MEDIATOR=mediator.pervoice.com
 CLIENT="./ebclient -s $MEDIATOR -p 4448 --timestamps -r"
+TEETS="$UNB ./eet-timestamped"
 for i in $(find ../../../SLTev/SLTev-cache/OStt-tt-files/ -type f -name "*_16K.wav");
 do
 	temp_name=$(basename $i)
@@ -47,8 +56,8 @@ do
 	TRDIR=${LOGSPATH}/${FROMASR}/${index_name}/${file_name}.transcript
     mkdir -p $DIR
     mkdir -p $TRDIR
-	echo "$UNB $CLIENT -f $FROMASR -i en-EU -t text < $i | $TEETS $DIR/${file_name}.en.asr" > $DIR/cmd
+	echo "$UNB $CLIENT -f $FROMASR -i en-EU -t text < $i | $TEETS $DIR/${file_name}.temp.asr" > $DIR/cmd
 	bash $DIR/cmd
-	cat "$DIR/${file_name}.en.asr" | cut -d" " -f2- | online-text-flow-events --text | sed -r '/^\s*$/d' > ${TRDIR}/${file_name}.en.asr
-	asrev_cmd="../../../SLTev/ASRev_updated/ASRev --ost $(basename "$temp_name" | rev | cut -d. -f2- | cut -d "_" -f2- | rev)t --asr ${TRDIR}/${file_name}.en.asr"
+
+	cat "$DIR/${file_name}.temp.asr" | cut -d" " -f2- | online-text-flow-events --text | sed -r '/^\s*$/d' > ${TRDIR}/${file_name}.en.asr
 done
